Creando los contant del componente HOME y sacando una clase hover general
