+++ src/App.css
@@ -5,6 +5,11 @@
   min-height: 100%;
 }
 
+.hover:hover {
+  transform: scale(1.05);
+  box-shadow: 0px 0px 10px 2px rgba(0, 0, 0, 0.5);
+}
+
 a {
   color: black;
   text-decoration: none;
