+++ src/pages/Home/components/index.css
@@ -10,11 +10,6 @@
     transition: all 0.2s ease-in-out;
 }
 
-.card:hover {
-    transform: scale(1.05);
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
-}
-
 .programCard {
     width: 80%;
     max-width: 950px;
@@ -39,12 +34,6 @@
     border-radius: 10px;
     box-shadow: 0px 0px 5px 1px rgba(0, 0, 0, 0.3);
     transition: all 0.2s ease-in-out;
-}
-
-.card iframe:hover,
-.card img:hover {
-    transform: scale(1.1);
-    box-shadow: 0px 0px 10px 2px rgba(0, 0, 0, 0.5);
 }
 
 /* Estilos para el contenido de la tarjeta */
